#4 Datatype 및 코드 수정

+++ src/main/java/com/core/toy3/src/like/model/request/LikeRequest.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class LikeRequest {
-    private Long memberId;
-    private Long travelId;
+    private Integer memberId;
+    private Integer travelId;
 }
